Move test fakes to owning modules
Each fake now lives in a `test/` subpackage of the module that owns the
interface it implements, instead of being bundled in a centralized
`feature/homescreen/test/Fakes.kt` file. This ensures test utilities
ship alongside the code they test and removes cross-module test coupling.
