+++ core/platform/src/commonMain/kotlin/space/be1ski/vibits/core/platform/test/FakeLocalConfigProvider.kt
@@ -1,0 +1,6 @@
+package space.be1ski.vibits.core.platform.test
+
+import space.be1ski.vibits.core.platform.env.LocalConfigProvider
+
+fun createFakeLocalConfigProvider(config: Map<String, String> = emptyMap()): LocalConfigProvider =
+  LocalConfigProvider { key -> config[key] }
