+++ feature/auth/data/src/commonMain/kotlin/space/be1ski/vibits/feature/auth/data/test/FakeCredentialsStore.kt
@@ -1,0 +1,20 @@
+package space.be1ski.vibits.feature.auth.data.test
+
+import space.be1ski.vibits.feature.auth.data.LocalCredentials
+import space.be1ski.vibits.feature.auth.data.platform.CredentialsStore
+
+class FakeCredentialsStore(
+  initial: LocalCredentials = LocalCredentials(baseUrl = "", token = ""),
+) : CredentialsStore {
+  var stored: LocalCredentials = initial
+    private set
+  var saveCalls: Int = 0
+    private set
+
+  override fun load(): LocalCredentials = stored
+
+  override fun save(credentials: LocalCredentials) {
+    stored = credentials
+    saveCalls += 1
+  }
+}
