+++ feature/habits/domain/build.gradle.kts
@@ -17,7 +17,6 @@ kotlin {
     }
     commonTest {
       dependencies {
-        implementation(projects.feature.homescreen)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
       }
