+++ feature/habits/presentation/build.gradle.kts
@@ -28,7 +28,6 @@ kotlin {
     commonTest {
       dependencies {
         implementation(projects.core.elm.test)
-        implementation(projects.feature.homescreen)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
       }
