+++ feature/memos/data/build.gradle.kts
@@ -33,7 +33,7 @@ kotlin {
 
     commonTest {
       dependencies {
-        implementation(projects.feature.homescreen)
+        implementation(projects.feature.sync.data)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
         implementation(libs.ktor.client.mock)
