+++ feature/memos/data/src/commonTest/kotlin/space/be1ski/vibits/feature/memos/data/OnlineMemosRepositoryTest.kt
@@ -13,9 +13,9 @@ import io.ktor.http.headersOf
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 import space.be1ski.vibits.feature.auth.domain.model.Credentials
-import space.be1ski.vibits.feature.homescreen.test.FakeCredentialsRepository
-import space.be1ski.vibits.feature.homescreen.test.FakeMemoCache
+import space.be1ski.vibits.feature.auth.domain.test.FakeCredentialsRepository
 import space.be1ski.vibits.feature.memos.data.remote.MemosApi
+import space.be1ski.vibits.feature.memos.data.test.FakeMemoCache
 import space.be1ski.vibits.feature.memos.domain.model.Memo
 import kotlin.test.Test
 import kotlin.test.assertEquals
