+++ feature/memos/presentation/src/commonTest/kotlin/space/be1ski/vibits/feature/memos/presentation/effect/MemosLoadEffectHandlerTest.kt
@@ -2,8 +2,8 @@ package space.be1ski.vibits.feature.memos.presentation.effect
 
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.test.runTest
-import space.be1ski.vibits.feature.homescreen.test.FakeMemosRepository
 import space.be1ski.vibits.feature.memos.domain.model.Memo
+import space.be1ski.vibits.feature.memos.domain.test.FakeMemosRepository
 import space.be1ski.vibits.feature.memos.domain.usecase.LoadCachedMemosUseCase
 import space.be1ski.vibits.feature.memos.domain.usecase.LoadMemosUseCase
 import space.be1ski.vibits.feature.memos.presentation.action.MemosAction
