+++ feature/mode/presentation/build.gradle.kts
@@ -26,7 +26,6 @@ kotlin {
     commonTest {
       dependencies {
         implementation(projects.core.elm.test)
-        implementation(projects.feature.homescreen)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
       }
