+++ feature/settings/domain/build.gradle.kts
@@ -14,7 +14,6 @@ kotlin {
     }
     commonTest {
       dependencies {
-        implementation(projects.feature.homescreen)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
       }
