+++ feature/settings/presentation/build.gradle.kts
@@ -29,7 +29,7 @@ kotlin {
     commonTest {
       dependencies {
         implementation(projects.core.elm.test)
-        implementation(projects.feature.homescreen)
+        implementation(projects.feature.onboarding.domain)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
       }
