+++ feature/settings/presentation/src/commonTest/kotlin/space/be1ski/vibits/feature/settings/presentation/SettingsEffectHandlerTest.kt
@@ -4,17 +4,17 @@ import kotlinx.coroutines.test.runTest
 import space.be1ski.vibits.core.platform.locale.AppLanguage
 import space.be1ski.vibits.core.platform.locale.LocaleProvider
 import space.be1ski.vibits.core.platform.mode.AppMode
+import space.be1ski.vibits.feature.auth.domain.test.FakeCredentialsRepository
 import space.be1ski.vibits.feature.auth.domain.usecase.SaveCredentialsUseCase
-import space.be1ski.vibits.feature.homescreen.test.FakeAppModeRepository
-import space.be1ski.vibits.feature.homescreen.test.FakeCredentialsRepository
-import space.be1ski.vibits.feature.homescreen.test.FakeMemoStorageManager
-import space.be1ski.vibits.feature.homescreen.test.FakeOnboardingStore
-import space.be1ski.vibits.feature.homescreen.test.FakePreferencesRepository
 import space.be1ski.vibits.feature.memos.domain.repository.ConnectionTester
+import space.be1ski.vibits.feature.memos.domain.test.FakeMemoStorageManager
+import space.be1ski.vibits.feature.mode.domain.test.FakeAppModeRepository
 import space.be1ski.vibits.feature.mode.domain.usecase.ResetAppUseCase
 import space.be1ski.vibits.feature.mode.domain.usecase.ResetAppWithMemosUseCase
 import space.be1ski.vibits.feature.mode.domain.usecase.SwitchAppModeUseCase
+import space.be1ski.vibits.feature.onboarding.domain.test.FakeOnboardingStore
 import space.be1ski.vibits.feature.settings.domain.model.AppTheme
+import space.be1ski.vibits.feature.settings.domain.test.FakePreferencesRepository
 import space.be1ski.vibits.feature.settings.domain.usecase.SaveLanguageUseCase
 import space.be1ski.vibits.feature.settings.domain.usecase.SaveThemeUseCase
 import space.be1ski.vibits.feature.settings.presentation.action.SettingsAction
