+++ feature/sync/data/build.gradle.kts
@@ -30,7 +30,6 @@ kotlin {
 
     commonTest {
       dependencies {
-        implementation(projects.feature.homescreen)
         implementation(kotlin("test"))
         implementation(libs.kotlinx.coroutines.test)
         implementation(libs.ktor.client.mock)
